레이어트 아키텍처를 적용한 설계 및 구현 domain - UserRequest, UserResponse  controller - UserRestController  service - UserService  repository - UserRepository

+++ src/main/java/com/example/springbootjpa/repository/UserRespository.java
@@ -3,5 +3,5 @@ package com.example.springbootjpa.repository;
 import com.example.springbootjpa.domain.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRespository extends JpaRepository<User,Integer> {
+public interface UserRespository extends JpaRepository<User,Long> {
 }
